Egor Buzanov

+++ src/main2.cpp
@@ -3,7 +3,7 @@
 int main()
 {
     const size_t arraySize = 10;
-    double *array = createArr<double, arraySize>(gen);
+    int *array = createArr<int, arraySize>(gen);
     for (size_t i = 0; i < arraySize; i++)
         std::cout << array[i] / 100 << " ";
     return 0;
